+++ 305.Application/Base/Response/Responses.cs
@@ -1,5 +1,5 @@
-﻿using _305.BuildingBlocks.Enums;
-using _305.BuildingBlocks.Text;
+﻿using _305.BuildingBlocks.Text;
+using _305.BuildingBlocks.Utils;
 using Serilog;
 
 namespace _305.Application.Base.Response;
